get price to display in bookshelf + search

+++ database/migrations/2023_01_22_220228_add_price_column_to_bookshelf_item.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('bookshelf_items', function (Blueprint $table) {
-            $table->integer('price')->after('type')->nullable();
+            $table->decimal('price', 8, 2)->after('type')->nullable();
         });
     }
